METAL-1484: Support using a block device as the base ISO
Calling Stat() on a block device always returns a size of 0, but other
than this there is no reason we cannot treat it as a normal file. Use
Seek() instead to find the size of a block device, and return an error
if the base file is neither a block device nor a regular file.

This allows us to use an attached Live ISO as the base file without
making a copy of it in RAM.